removed-text-edited-headings
I removed the copy: "EEW Report Cards analyze and present data on violations of environmental laws, inspections of polluting facilities, and enforcement actions by the EPA. So far, we’ve generated reports for the congressional districts of each Representative on the House Energy and Commerce Committee (E&C), and for the state of each Senator on the Environment and Public Works Committee (EPW). These two committees are responsible for oversight of the EPA, an agency with a congressional mandate to enforce environmental laws. Our goal is to shed light on the chronic and routine violations of environmental laws and the failure of EPA to enforce them. "
...and will move it to below the interactive graphic.

I also played with the hyperlinked report heading (and added the hyperlink); I tried to make it a heading and center it rather then just have it bolded. I also centered the "select and view completed report cards by representative below:" and made it bolded instead of a header. @Frijol I can't tell exactly how it will look on the finished site so let me know if it looks weird